+++ tests/Unit/Util/EnvTest.php
@@ -26,24 +26,24 @@
 namespace LibreNMS\Tests\Unit\Util;
 
 use LibreNMS\Tests\TestCase;
-use LibreNMS\Util\Env;
+use LibreNMS\Util\EnvHelper;
 
 class EnvTest extends TestCase
 {
     public function testParseArray()
     {
         putenv('PARSETEST=one,two');
-        $this->assertSame(['one', 'two'], Env::parseArray('PARSETEST'), 'Could not parse simple array');
-        $this->assertSame(['default'], Env::parseArray('PARSETESTNOTSET', 'default'), 'Did not get default value as expected');
-        $this->assertSame(null, Env::parseArray('PARSETESTNOTSET'), 'Did not get null as expected when env not set');
-        $this->assertSame(3, Env::parseArray('PARSETESTNOTSET', 3), 'Did not get default value (non-array) as expected');
-        $this->assertSame('default', Env::parseArray('PARSETESTNOTSET', 'default', ['default']), 'Did not get default value as expected, excluding it from exploding');
+        $this->assertSame(['one', 'two'], EnvHelper::parseArray('PARSETEST'), 'Could not parse simple array');
+        $this->assertSame(['default'], EnvHelper::parseArray('PARSETESTNOTSET', 'default'), 'Did not get default value as expected');
+        $this->assertSame(null, EnvHelper::parseArray('PARSETESTNOTSET'), 'Did not get null as expected when env not set');
+        $this->assertSame(3, EnvHelper::parseArray('PARSETESTNOTSET', 3), 'Did not get default value (non-array) as expected');
+        $this->assertSame('default', EnvHelper::parseArray('PARSETESTNOTSET', 'default', ['default']), 'Did not get default value as expected, excluding it from exploding');
 
         putenv('PARSETEST=');
-        $this->assertSame([''], Env::parseArray('PARSETEST', null, []), 'Did not get empty string as expected when env set to empty');
+        $this->assertSame([''], EnvHelper::parseArray('PARSETEST', null, []), 'Did not get empty string as expected when env set to empty');
 
         putenv('PARSETEST=*');
-        $this->assertSame('*', Env::parseArray('PARSETEST', null, ['*', '*']), 'Did not properly ignore exclude values');
+        $this->assertSame('*', EnvHelper::parseArray('PARSETEST', null, ['*', '*']), 'Did not properly ignore exclude values');
 
         // clean the environment
         putenv('PARSETEST');
